+++ test.py
@@ -1,0 +1,7 @@
+import os
+
+my_icons_dir = os.path.join(os.path.dirname(__file__), "icons")
+
+my_icons_dir = str(my_icons_dir)
+
+print(my_icons_dir)