+++ docs/docs/addon-information/changelog.md
@@ -114,3 +114,18 @@ STABLE
 ### Fixed
 - Fixed [Incorrect context error](https://github.com/MattAshpole/BlenderRizomUVBridge/issues/5)
 - Changed UV map matching for multiple UV sets from map index to map name to avoid a bug where the index would change causing UVs to be transferred to the wrong UV set.
+
+
+## Version 0.3.2
+
+{: .d-inline-block }
+IN DEVELOPMENT
+{: .label .label-yellow }
+
+### Improved
+
+- Changed all settings to addon level preferences so that they are remembered between Blender sessions.
+
+### Added
+
+- Button to reset settings to default for each panel.